requirements.txt
Flask
pytesseract
Pillow
werkzeug
opencv-python
numpy
